# usoft-backend/src/docs/openapi-usof.yaml
openapi: 3.0.3
info:
  title: USOF Backend API
  version: 1.2.0
  description: |
    Forum platform backend (No ORM, pure SQL + mysql2). This spec enumerates all implemented routes, parameters, request/response schemas, and auth rules.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id: { type: integer }
        login: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        emailConfirmed: { type: boolean }
        avatarUrl: { type: string, nullable: true }
        role: { type: string, enum: [user, admin] }
        rating: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, login, fullName, email, role, rating, createdAt, updatedAt]
    UserCreate:
      type: object
      properties:
        login: { type: string, minLength: 3, maxLength: 50 }
        password: { type: string, minLength: 6, maxLength: 100 }
        passwordConfirm: { type: string }
        fullName: { type: string, minLength: 3, maxLength: 100 }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
      required: [login, password, passwordConfirm, fullName, email, role]
    UserUpdate:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
...
    Post:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        status: { type: string, enum: [active, inactive] }
        likesCount: { type: integer }
        dislikesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        authorId: { type: integer }
        authorLogin: { type: string }
        authorFullName: { type: string }
        authorRating: { type: integer }
      required: [id, title, content, status, likesCount, dislikesCount, createdAt, updatedAt]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List users (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Users]
      summary: Create user (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/users/avatar:
    patch:
      tags: [Users]
      summary: Upload/replace my avatar
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required: [avatar]
      responses:
        '200':
          description: Avatar URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatarUrl: { type: string }
                required: [avatarUrl]
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id (self or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Users]
      summary: Update user (self or admin; only admin may change role)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Users]
      summary: Delete user (self or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/users/{id}/favorites:
    get:
      tags: [Favorites, Users]
      summary: List a user's favorite posts (self or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Favorite posts visible to the requester
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Post' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ===================== Posts ======================
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Admin]
      summary: Create user (emailConfirmed=1)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, email, fullName, password]
              properties:
                login: { type: string }
                email: { type: string, format: email }
                fullName: { type: string }
                password: { type: string }
                role: { type: string, enum: [user, admin], default: user }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      emailConfirmed: { type: boolean }
  /api/admin/users/{id}:
    patch:
      tags: [Admin]
      summary: Update user fields (fullName/email/role)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string }
                email: { type: string, format: email }
                role: { type: string, enum: [user, admin] }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Delete user
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
...